refactor: :recycle: simplify main.ts
