+++ githubCrawler/crawler/sites.json
@@ -1,0 +1,21 @@
+{
+  "github/trendings": {
+    "selector": "h2 a.Link |> prepend:https://github.com",
+    "pages": [
+      "https://github.com/trending/typescript?since=monthly",
+      "https://github.com/trending/javascript?since=monthly",
+      "https://github.com/trending/php?since=monthly",
+      "https://github.com/trending/python?since=monthly",
+      "https://github.com/trending/rust?since=monthly"
+    ]
+  },
+  "codebase.show": {
+    "selector": "a[href^='https://github.com'] > div |> parent",
+    "pages": [
+      "https://codebase.show/projects/realworld?category=frontend",
+      "https://codebase.show/projects/realworld?category=backend",
+      "https://codebase.show/projects/realworld?category=fullstack",
+      "https://codebase.show/projects/todomvc"
+    ]
+  }
+}
